+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix not being able to call global functions without arguments
 - Fix multiple inheritance not rendering blocks as expected for nested blocks
+- Allow filters on key/value for loop containers
 
 ## 0.10.5 (2017-05-13)